Split database and middleware functions into modules.
TODO: redo initializing sequence using new organization system.
